Move install tracking to its own file
