+++ src/components/services/Services.js
@@ -19,9 +19,9 @@ import { rootReducer } from '../../reducers/RootReducer.js'
 import { type Action } from '../../types/reduxActions.js'
 import { type State } from '../../types/reduxTypes.js'
 import errorAlert from '../../util/errorAlert.js'
+import { loadInstallReason } from '../../util/installReason.js'
 import loginStatusChecker from '../../util/loginStatusChecker.js'
 import perfLogger from '../../util/perfLogger.js'
-import { loadInstallReason } from '../../util/tracking.js'
 import { ModalProvider } from '../common/ModalProvider.js'
 import { Airship } from './AirshipInstance.js'
 import { AutoLogout } from './AutoLogout.js'
